requirements.txt
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
mysql-connector-python==8.0.28
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.0
protobuf==3.19.3
pycparser==2.21
python-dotenv==0.19.2
Wand==0.6.7
